requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0
